añadida cosa de los permisos

+++ commands/utility/delete.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder } = require("discord.js");
+const { SlashCommandBuilder, PermissionFlagsBits } = require("discord.js");
 const prisma = require("../../prisma/prisma.js");
 
 module.exports = {
@@ -16,30 +16,71 @@ module.exports = {
     async execute(interaction) {
         const aliasborrar = interaction.options.getString("alias");
 
-        await interaction.reply(
-            `Deleting coordinates from ${aliasborrar}...`,
-        );
-        try {
-            const result = await prisma.cords.deleteMany({
-                where: {
-                    guildId: interaction.guildId,
-                    alias: aliasborrar,
-                }})
-            if (result.count === 0){
-                await interaction.editReply(
-                    `No coordinates under "${aliasborrar}".`
+        const db = await prisma.derole.findFirst({
+            where: {
+                guildId: interaction.guildId,
+            }
+        });
+
+        let clearance = false;
+
+        const deletorRole = db?.DELETOR;
+
+        if ((deletorRole === "admin")){
+            if(interaction.member.permissions.has(PermissionFlagsBits.Administrator)){
+                clearance = true
+            }
+            else {
+                clearance = false
+                await interaction.reply(
+                {content: `You don't have permission to do this!`,
+                ephemeral: true,}
                 );
                 return;
             }
-            await interaction.editReply(
-                `"${aliasborrar}" deleted.`
+        } else if (deletorRole === "custom") {
+            const customRole = interaction.member.guild.roles.cache.find(role => role.name === "Nauta Admin");
+            if (customRole && interaction.member.roles.cache.has(customRole.id)){
+                clearance = true;
+            } else {
+            clearance = false;
+            await interaction.reply(
+                {content: `You don't have permission to do this!`,
+                ephemeral: true,}
             );
-        }
-        catch (error) {
-            console.error('Error: ', error)
-            await interaction.editReply(
-                `Unable to delete coordinates.`,
+            return;
+            }
+        } else {clearance = true}
+        
+        
+        if (clearance){
+            await interaction.reply(
+                {content: `Deleting coordinates from ${aliasborrar}...`,
+                ephemeral: true,}
             );
+            try {
+                const result = await prisma.cords.deleteMany({
+                    where: {
+                        guildId: interaction.guildId,
+                        alias: aliasborrar,
+                    }})
+                if (result.count === 0){
+                    await interaction.editReply(
+                        `No coordinates under "${aliasborrar}".`
+                    );
+                    return;
+                }
+                await interaction.editReply(
+                    `"${aliasborrar}" deleted.`
+                );
+            }
+            catch (error) {
+                console.error('Error: ', error)
+                await interaction.editReply(
+                    {content: `Unable to delete coordinates.`,
+                    ephemeral: true}
+                );
+            }
         }
     },
 };
